pyproject.toml
[project]
name = "python-yaml-validator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "dotenv>=0.9.9",
    "pytest>=9.0.2",
    "pyyaml>=6.0.3",
    "typer>=0.21.1",
]

[tool.pytest.ini_options]
pythonpath = ["."]
